requirements.txt
selenium
webdriver-manager
requests
python-dotenv
browserstack-sdk
pytest-selenium